+++ packages/wrapper-react/CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to npm module [@typefox/monaco-editor-react](https://www.npm
 - Support all arguments for monaco-vscode-api `initialize` [#756](https://github.com/TypeFox/monaco-languageclient/pull/756)
   - On startup the current containerRef is passed to the interal editor app/monaco-editor
 - Update to monaco-vscode-api 9.0.x [#749](https://github.com/TypeFox/monaco-languageclient/pull/749)
-  - Updated to `monaco-editor-wrapper@9.0.0-next.2` and `monaco-languageclient@6.0.0-next.2`. Updated all `@codingame/monaco-vscode` packages to `9.0.3`.
+  - Updated to `monaco-editor-wrapper@9.0.0-next.2` and `monaco-languageclient@6.0.0-next.2`. Updated all `@codingame/monaco-vscode` packages to `10.0.1`.
   - Enhancements to logging
 - monaco-languageclient config improvement, wrapper+languageclientwrapper improvements [#741](https://github.com/TypeFox/monaco-languageclient/pull/741)
 - Turned react component from class to function [#739](https://github.com/TypeFox/monaco-languageclient/pull/739)