+++ packages/wrapper/CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to npm module [monaco-editor-wrapper](https://www.npmjs.com/
 - Support all arguments for monaco-vscode-api `initialize` [#756](https://github.com/TypeFox/monaco-languageclient/pull/756)
   - This also allows to configure editor-, view- or workspace-service. This is a preparation for further enhancements.
 - Update to monaco-vscode-api 9.0.x [#749](https://github.com/TypeFox/monaco-languageclient/pull/749)
-  - Updated to `monaco-languageclient@9.0.0-next.2`. Updated all `@codingame/monaco-vscode` packages to `9.0.3`.
+  - Updated to `monaco-languageclient@9.0.0-next.2`. Updated all `@codingame/monaco-vscode` packages to `10.0.1`.
   - Enhancements to logging, worker factory and start order. The worker factory only accepts direct worker loading instructions from now on
 - monaco-languageclient config improvement, wrapper+languageclientwrapper improvements [#741](https://github.com/TypeFox/monaco-languageclient/pull/741)
   - Allows to configure more than one language client